// wetty-chat-mobile/src/generated/models/messagePreview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wetty Chat API
 * Real-time chat application backend API supporting groups, messaging, threads, stickers, invites, and push notifications.
 * OpenAPI spec version: 0.1.0
 */
import type { MentionInfo } from './mentionInfo';
import type { MessagePreviewAttachment } from './messagePreviewAttachment';
import type { MessagePreviewSticker } from './messagePreviewSticker';
import type { MessageType } from './messageType';
import type { SnowflakeID } from './snowflakeID';
import type { User } from './user';

export interface MessagePreview {
  attachments: MessagePreviewAttachment[];
  clientGeneratedId: string;
  createdAt: string;
  id: SnowflakeID;
  isDeleted: boolean;
  mentions: MentionInfo[];
  message?: string;
  messageType: MessageType;
  sender: User;
  sticker?: MessagePreviewSticker;
}